requirements.txt
numpy
pandas
opencv-python
terminaltables
Pillow
matplotlib
blessed
jupyterlab